Valida que el evento pertenezca al chefId webApi.

+++ WebApi/Controllers/ChefController.cs
@@ -1,6 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Services.Interfaces;
 using System;
+using System.Net;
+using System.Net.Http;
+using System.Web.Http;
 
 namespace WebApi.Controllers
 {
@@ -14,16 +17,27 @@ namespace WebApi.Controllers
             _eventService = eventService;
         }
 
-        [HttpGet]
-        [Route("/chef/cancel")]
-        public void Index(int eventId, int chefId)
+        [Microsoft.AspNetCore.Mvc.HttpGet]
+        [Microsoft.AspNetCore.Mvc.Route("/chef/cancel")]
+        public IActionResult Index(int eventId, int chefId)
         {
             if (chefId == 0)
             {
                 throw new ArgumentException("Inválido");
             }
 
-            _eventService.Delete(eventId, chefId);
+            if(_eventService.IsEventBelongToUser(eventId, chefId))
+            {
+                _eventService.Delete(eventId, chefId);
+                return Ok();
+            }
+            else
+            {
+                var message = string.Format("El id del evento no pertenece al id del usuario", eventId, chefId);
+                HttpError err = new HttpError(message);
+                return StatusCode(500, err);
+            }
+
         }
